Modify and expand type lookups.
Originally shcd_mapper was used to map device shcd names to common
names.  This change generalizes the mapper to include shcd and lldp
types.  The change also includes the common/shasta name into the
lookup list for consistency and to allow names like sw-spine01 to map
to the common name of sw-spine-001.