Remove NetworkNodeFactory as a singleton
Initially NetworkNodeFactory was created as a singleton to protect
unique node numbering.  Functionally, as it turns out, making the
factory as a singleton is not necessary and created side effects when
two models are created for comparison in the same runtime - e.g.
LLDP compared to SHCD.  In this case the factory as a singleton did not
create indepenent numbering - which is a problem.  This change removes
the singleton in the factory.