Add major feature to track ports and slots.
Previously canu would track logical connections between devices, but
was not able to map this to physical ports on the device.  This new
feature adds the concept of device physical ports which can be
subdivided into blocks of ports called slots.  Slots fit naturally in
server devices where slots like "ocp" and "pcie-slot1" create a real,
physical separation of ports.  The slot defaults to None if not
specified in the model.  Slots and ports allow tracking and enforcement
of physical connections in the model.

The change in the model maintains supports previous behavior where a
connection between devices is specified, but a port assingment is not
specified.  In this case the new model assigns ports automatically based
on architectural rules and standards, with minimal heuristics involved.

Exception handling has also been cleaned up and standardized.