Add port and slot tracking to cabling (LLDP).
This addition finishes the port and slot tracking plumbing in cabling
and shcd_cabling to track ports and slots.  The change also required a
small refactoring in port assignment during connections.

The strict flag was added to node connect to handle the following cases:
1. If an SHCD specifies that a port is used multiple times the
connection should not be allowed and an error should be returned.
This is new behavior specified by strict=True.
2. If LLDP data aggregated across all switches specifies a connection
multiple times, this is allowed.  The previous connection is detected,
the new connection logged as a warning and processing continues. Both
cabling and cabling_shcd use this behavior to prevent the need for
sorting of lists.