Add a HmnRow class for use with validate shcd
This commit adds a new class, which reflects what we expect in a row in the HMN
tab.  By default, the class doesn't deal with missing or incompatible columns.
The class presumes the happy path, where it recieves a row with the correct
columns.  The user can override this, as I do in this same commit.

Eventually, I'd like to see us just ingest the correct rows instead of dealing
with all the cray roadblocks, so this can set us up for that.

The class has simple attributes for things like:

- `hmnrow.source_port`
- `hmnrow.dest_loc`
- `hmnrow.required_headers`

Signed-off-by: Jacob Salmela <jacob.salmela@hpe.com>