Fix conflict between nornir and dataclasses
dataclasses 0.8 is getting pulled in implicitly by ... something ... in the context of a docker build.  This causes the program to not operate correctly as reported in https://github.com/Cray-HPE/canu/issues/78.  One possible fix is to simply constrain the version of dataclasses to 0.7 to prevent this anomalous behavior, another would be to identify what is pulling in dataclasses and address it more explicitly elsewhere.  Being unfamiliar with this package, I've chosen the former.