44 Build Once and Use Git-Tags
This removes the .version files and uses git-tags. This also builds canu
inside of the RPM spec file, and then uses the RPM from there.

The spec file will create a canu user.

Since the Docker image is still Debian, this does not incorporate the
RPM into the image build.