Relates to #44 - Fixes stable RPM's
Stable builds of `canu` are reporting a version string with `dev1`
appended to the end, this implies the build was done with a dirty repo
(uncommitted changes). This occurs because the CSM metadata we add to
the spec file causes the repo to become dirty. This change tells Git to
assume that the `.spec` file is clean, thus preventing git from
identifying the repo as dirty during the build.