Prevent `python2` from being invoked
The `python-rpm-macros` in the  SP3 build can occaisionally attempt building `python2`, this is random
and causes a build failure. We don't care about Python2; tell the SPEC
to build only for Python3.

Also remove the erroneous `%define` and `%global` that are no longer
used, and add `BuildRequires` to explicitly note we're using
`python-rpm-*` packages.

Run the RPM build as root, making the installed binary for `pyinstaller`
available in the path. I wasn't able to figure out how to amend the
`PATH` available to the RPM, and by removing the hardocde to Jenkins in
the `.spec` I needed to update the `PATH` or run as `root`.