Shrink `site-lib` footprint
The automatic finding of packages by `setuptools` was including `*`
(everything) in this repository in the `site-lib` folder. This bloats
local installations of `canu`.

Shrink the size by including only the code we need, and what's defined
in `MANIFEST.in`.

Also pin the python-rpm-macros to `python3.10`

Build by python version instead of sles

Disable PR-merge build to lighten the load on Jenkins.