Versioning information and pre-releases (#267)
* Versioning information and pre-releases

This updates the releasing procedure, including directions for
publishing a pre-release (alpha, beta, release candidate).

The `Jenkinsfile.github` declaritive pipeline now only publishes final
release tags to stable, pre-releases still publish to unstable.

This also cleans up the `README.md`, fixing the changelog links and
table of contents.

The "Versioning" section now elaborates on pre-releases, and the
`setuptools_scm` configuration now guesses the next version.

* Prevent `python2` from being invoked

The `python-rpm-macros` in the  SP3 build can occaisionally attempt building `python2`, this is random
and causes a build failure. We don't care about Python2; tell the SPEC
to build only for Python3.

Also remove the erroneous `%define` and `%global` that are no longer
used, and add `BuildRequires` to explicitly note we're using
`python-rpm-*` packages.

Run the RPM build as root, making the installed binary for `pyinstaller`
available in the path. I wasn't able to figure out how to amend the
`PATH` available to the RPM, and by removing the hardocde to Jenkins in
the `.spec` I needed to update the `PATH` or run as `root`.

* Language fixups

* Shrink `site-lib` footprint

The automatic finding of packages by `setuptools` was including `*`
(everything) in this repository in the `site-lib` folder. This bloats
local installations of `canu`.

Shrink the size by including only the code we need, and what's defined
in `MANIFEST.in`.

Also pin the python-rpm-macros to `python3.10`

Build by python version instead of sles

Disable PR-merge build to lighten the load on Jenkins.

* Dep fix