Stable Build Needs
Fixes Docker-in-Docker publishing; since a Docker container is used to
provide `os-release` for determining where to publish the RPMs, the
ClamAV `Docker` container needs `docker.sock` mounted so it can run
Docker-in-Docker.

Additionally, the SSH public keys on the node are not available to the Docker container for
sending RPMs to DST for signing. This mounts the SSH keys for the Docker
container to successfully sign the RPM.