Optimize Jenkins, Build multi-arch Docker Image, and Tidy `Makefile` (#355)
* Simplify build resources

This reduces the number of resources needed for the canu build by using
_one_ host that spins up multiple docker containers instead of multiple
hosts with their own docker containers.

* Enable linux/arm64 Docker Images, prep RPM

Broaden Docker images to support arm64, and simplify some of the image
calls using newer `Makefile` syntax.

This also preps the RPM build with `--target ${ARCH}`, meaning if/when
we need an `aarch64` `canu.rpm` all we have to do is update the
`Jenkinsfile` (theoretically).

* Neaten up `Makefile`

* Remove Py3.9 and enable RPM rebuilds

* Try a large builder with both images

* `Dockerfile` tweaks - prod no longer depends on dev

This change shaves off some build time.

prod depends on prod_build, which uses a `--from dev` call. This means building prod will also build dev, which adds ~10min to the prod image build unless the dev image is already sitting around.

Secondly, the `COPY` statement is only omitting what looks like two files; `README.md`, and `Jenkinsfile.github`. The `README.md` is technically desired by the `pyproject.toml` file, so it seems harmless to add. The `Jenkinsfile.github` should still be omitted and now is included in the `.dockerignore`, as well as `canu.spec` (which is no longer listed in the `COPY`. The `COPY` is moved into the `deps` layer, so all targets inheriting from `deps` have the files by default.

Thirdly, I kept running into `/bin/bash does not exist` when becoming the `canu` user, so I added `bash` to the list of packages to be installed.

Fourthly, this replaces the `nox -e docs` with the actual `sphinx` call so the previously installed `.[docs]` dependencies can be used instead of re-installed into a nox venv (saving more time).

Lastly this consolidates a few `ENV` lines.