add integration tests using shellspec
this adds a spec/ folder, where all the tests live.
To begin, I just added simple tests for the --help output of every
command.  More work is needed to run more tests and additional mocks are
needed for the majority of the commands.

This gives us a good start and works as a nice sanity check.

Signed-off-by: Jacob Salmela <jacob.salmela@hpe.com>