Add post-release support
Enable post-release tags. If a tag with `.postN` (where `N` is an
integer) is created, consider this tag as a stable rebuild of a previous
release.

When a post-release tag is created, the created RPMs will bump their
`%{RELEASE}` value to allow package managers to properly install the
repackaged RPM if the previously packaged RPM is present. Since the
`%{VERSION}` is the same, this will convey to the user that the
developer is shipping the same release of software (same content) but
either documentation or packaging changes were brought in.

The advantage of a post-release is it does not look as suspicious as a
version bump looks. E.g. 1.0.0 vs 1.0.1 raises suspicion that content
(bugfixes/patches/code-change) has been pulled in. Whereas a
post-release uses the same version string to indicate the content is
exactly the same.

NOTE: Since images use layers, the image tags will never indicate
anything differently when a post-release is taken. The tags will
overwrite each other in the registry, and users pulling the image will
simply notice new layers being pulled. Post-releases are confusing in
the image tag context, so they've been left out.