Enable manual re-runs and sleep longer
Today a build matrix finished its last build 10sec after the others and
its RPMs didn't get appended to the release. The RPMs then had to be
manually added to the release.

This adds a re-run ability to each of our promotion workflows, to enable
re-running and updating an already published release. Additionally this
adds a few seconds to the wait periods to help catch the edge cases.