update dependencies
this updates quite a bit.  it brings python up to 3.12, alpine to edge
in order to aid in the remediation of CVE-2024-6387.  at the time of
this commit, the patch is only available in edge.  as a result of this,
our pipeline marks this as unstable because it isn't a published
release.

beyond the CVE, I opted to bump everything since it has been nearly a
year and the tests all still pass, minus a few that needed a slight
adjustment for the new python version.

pinning the versions so strictly can be helpful, but difficult at the
same time.  in this case--as well as times in the past--it has been hard
to bump a single version of a dependency when a CVE or update is needed.
it usually triggers a cascade of other depdency warnings.

we also have a weird mix of dependencies from both the normal python
install and the container version of canu.  we had planned to drop the
binary version and bundle the image with the rpm, but that was more than
a year ago, and no progress has been made towards that.

to that end, I removed all the pinned versions and just went with the
latest.  since we do not release that often, we will be able to catch
problems when building locally and we can choose to pin at the time if
need be, but the CVE in this commit is not the only one out there, so
bringing everything up to latest (at least for now) seems to be the best
option IMO.

Signed-off-by: Jacob Salmela <jacob.salmela@hpe.com>